third_party/blink/web_tests/external/wpt/css/css-masking/parsing/clip-path-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: getComputedStyle().clipPath</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes/#clip-path-property">
<meta name="assert" content="clip-path computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<style>
  #target {
    font-size: 40px;
  }
</style>
<div id="target"></div>
<script>
test_computed_value("clip-path", "circle(calc(10px + 0.5em) at -50% 50%)", "circle(30px at -50% 50%)");
test_computed_value("clip-path", "circle(calc(10px - 0.5em) at 50% -50%)", "circle(0px at 50% -50%)");
test_computed_value("clip-path", "ellipse(at 50% 50%)");
test_computed_value("clip-path", "ellipse(60% closest-side at 50% 50%)");

test_computed_value("clip-path", "xywh(0 1% 2px 0.5em)", "xywh(0px 1% 2px 20px)");
test_computed_value("clip-path", "xywh(0px 1% 2px 0.5em round 0)", "xywh(0px 1% 2px 20px)");
test_computed_value("clip-path", "xywh(0px 1% 2px 3% round 0 1px)", "xywh(0px 1% 2px 3% round 0px 1px)");
test_computed_value("clip-path", "xywh(0px 1% 2px 3% round 0px 1px 2em)", "xywh(0px 1% 2px 3% round 0px 1px 80px)");
test_computed_value("clip-path", "xywh(0px 1% 2px 3% round 0px 1px 2% 3px)");
</script>
</body>
</html>
