third_party/blink/web_tests/external/wpt/css/css-masking/parsing/clip-path-computed-expected.txt

This is a testharness.js-based test.
PASS Property clip-path value 'circle(calc(10px + 0.5em) at -50% 50%)'
PASS Property clip-path value 'circle(calc(10px - 0.5em) at 50% -50%)'
PASS Property clip-path value 'ellipse(at 50% 50%)'
PASS Property clip-path value 'ellipse(60% closest-side at 50% 50%)'
FAIL Property clip-path value 'xywh(0 1% 2px 0.5em)' assert_equals: expected "inset(1% calc(100% - 2px) calc(99% - 20px) 0px)" but got "rect(0px 1% 2px 20px)"
FAIL Property clip-path value 'xywh(0px 1% 2px 0.5em round 0)' assert_equals: expected "inset(1% calc(100% - 2px) calc(99% - 20px) 0px)" but got "rect(0px 1% 2px 20px)"
FAIL Property clip-path value 'xywh(0px 1% 2px 3% round 0 1px)' assert_equals: expected "inset(1% calc(100% - 2px) 96% 0px round 0px 1px)" but got "rect(0px 1% 2px 3% round 0px 1px)"
FAIL Property clip-path value 'xywh(0px 1% 2px 3% round 0px 1px 2em)' assert_equals: expected "inset(1% calc(100% - 2px) 96% 0px round 0px 1px 80px)" but got "rect(0px 1% 2px 3% round 0px 1px 80px)"
FAIL Property clip-path value 'xywh(0px 1% 2px 3% round 0px 1px 2% 3px)' assert_equals: expected "inset(1% calc(100% - 2px) 96% 0px round 0px 1px 2% 3px)" but got "rect(0px 1% 2px 3% round 0px 1px 2% 3px)"
FAIL Property clip-path value 'rect(auto auto auto auto)' assert_equals: expected "inset(0%)" but got "rect(auto auto auto auto)"
FAIL Property clip-path value 'rect(0 1% 2px 0.5em)' assert_equals: expected "inset(0px 99% calc(100% - 2px) 20px)" but got "rect(0px 1% 2px 20px)"
FAIL Property clip-path value 'rect(0px 1% auto 0.5em round 0)' assert_equals: expected "inset(0px 99% 0% 20px)" but got "rect(0px 1% auto 20px)"
FAIL Property clip-path value 'rect(0px 1% auto 3% round 0 1px)' assert_equals: expected "inset(0px 99% 0% 3% round 0px 1px)" but got "rect(0px 1% auto 3% round 0px 1px)"
FAIL Property clip-path value 'rect(0px 1% auto 3% round 0px 1px 2em)' assert_equals: expected "inset(0px 99% 0% 3% round 0px 1px 80px)" but got "rect(0px 1% auto 3% round 0px 1px 80px)"
FAIL Property clip-path value 'rect(0px 1% auto 3% round 0px 1px 2% 3px)' assert_equals: expected "inset(0px 99% 0% 3% round 0px 1px 2% 3px)" but got "rect(0px 1% auto 3% round 0px 1px 2% 3px)"
Harness: the test ran to completion.
